// dashboard/src/api/models/workflowNodeSchema.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Scheduler Public API (v1)
 * OpenAPI spec version: 1.3.0
 */
import type { WorkflowNodeSchemaParameters } from './workflowNodeSchemaParameters';
import type { WorkflowNodeSchemaResults } from './workflowNodeSchemaResults';

/**
 * JSON Schemas for parameters/results.
 */
export type WorkflowNodeSchema = {
  parameters?: WorkflowNodeSchemaParameters;
  results?: WorkflowNodeSchemaResults;
};